app/generated_project/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==1.4.41
passlib==1.7.4
python-jose==3.3.0
